requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.8
django-crispy-forms==1.9.1
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
joblib==0.16.0
numpy==1.19.0
pandas==1.0.5
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
whitenoise==5.1.0
